refactor(userspace-dp): #6386 extract flowless_verdict sibling
Move the flowless (no-L4) LocalDelivery verdict helpers out of
poll_descriptor/mod.rs into a new pub(super) sibling flowless_verdict.rs:
enum FlowlessLocalVerdict, ipv6_ext_header_over_limit_drop,
flowless_local_delivery_verdict, flowless_base_resolution, plus the two
stranded test modules ipv6_ext_header_drop_tests and
flowless_local_delivery_tests (loaded from the sibling via #[path]).
Each moved bare fn/enum becomes pub(super); nothing widens further.

Pure code motion. Bodies byte-identical to their prior location
(verified with git diff --color-moved). The three per-packet-capable
fns GAIN #[inline] (the only non-motion change — restores same-CGU
inlining eligibility across the new module boundary per the #6386
hot-path preservation contract); the enum is attr-verbatim.

Cross-module edges made explicit: flowless_verdict imports
emit_pending_filter_log + host_inbound_gated_lo0_action from filter and
emit_host_inbound_deny + emit_junos_host_deny + junos_host_policy_eval
from host_inbound_policy. With flowless_local_delivery_verdict gone from
mod.rs, its now-dead imports are pruned: junos_host_policy_eval +
emit_junos_host_deny from host_inbound_policy, and emit_pending_filter_log
from filter (host_inbound_gated_lo0_action stays — the loop still uses
it). mod.rs re-imports the four moved names via
`use flowless_verdict::{...}` so every loop call site is textually
unchanged.

The two test modules move byte-identical inside a nested #[path] tests
sibling (a file-top `use super::*;` re-glob wires the established
4-level glob chain; the module bodies are unchanged, indentation and
all).

Validation: cargo build (release) clean; cargo test --no-run clean;
moved bodies byte-identical modulo pub(super) and the three #[inline].
